Create user settings form component
