+++ SnapWar/ViewControllers/SWAccountViewController.swift
@@ -4,13 +4,11 @@ import SCSDKBitmojiKit
 class SWAccountViewController: UIViewController {
     let displayName = UILabel()
     let avatar = SWAvatar()
-    let settingsColumn = UIStackView()
-    let notificationsSwitch = SWSwitch()
-    
+    let settingsForm = SWUserSettingsForm()
     
     override func viewDidLoad() {
         super.viewDidLoad()
-         setup()
+        setup()
     }
     
     override func viewWillAppear(_ animated: Bool) {
@@ -23,9 +21,7 @@ class SWAccountViewController: UIViewController {
         view?.backgroundColor = SWColorType.white
         view.addSubview(avatar)
         view.addSubview(displayName)
-        view.addSubview(settingsColumn)
-
-        setupSettings()
+        view.addSubview(settingsForm)
         
         displayName.text = SWUserEntity.instance.displayName
         displayName.textColor = SWColorType.black
@@ -41,40 +37,9 @@ class SWAccountViewController: UIViewController {
             avatar.heightAnchor.constraint(equalToConstant: 150),
             avatar.bottomAnchor.constraint(equalTo: displayName.topAnchor, constant: -10),
             avatar.centerXAnchor.constraint(equalTo: view.centerXAnchor),
-            settingsColumn.topAnchor.constraint(equalTo: displayName.bottomAnchor, constant: 10),
-            settingsColumn.leadingAnchor.constraint(equalTo: view.leadingAnchor),
-            settingsColumn.trailingAnchor.constraint(equalTo: view.trailingAnchor),
+            settingsForm.topAnchor.constraint(equalTo: displayName.bottomAnchor, constant: 10),
+            settingsForm.leadingAnchor.constraint(equalTo: view.leadingAnchor),
+            settingsForm.trailingAnchor.constraint(equalTo: view.trailingAnchor),
         ])
-    }
-    
-    fileprivate func setupSettings() {
-        settingsColumn.axis = .vertical
-        settingsColumn.alignment = .fill
-        settingsColumn.distribution = .equalCentering
-        settingsColumn.translatesAutoresizingMaskIntoConstraints = false
-        
-        let notificationRow = createFieldRow(label: "Allow notifications")
-        notificationRow.addArrangedSubview(notificationsSwitch)
-        notificationRow.translatesAutoresizingMaskIntoConstraints = false
-        
-        settingsColumn.addArrangedSubview(notificationRow)
-        
-    }
-    
-    fileprivate func createFieldRow(label labelText: String) -> UIStackView {
-        let row = UIStackView();
-        row.axis = .horizontal
-        row.alignment = .center
-        row.distribution = .equalSpacing
-        row.layoutMargins = UIEdgeInsets(top: 10, left: 20, bottom: 10, right: 20)
-        row.isLayoutMarginsRelativeArrangement = true
-        row.translatesAutoresizingMaskIntoConstraints = false
-        
-
-        let label = UILabel()
-        label.text = labelText
-        row.addArrangedSubview(label)
-        
-        return row;
     }
 }
